# Use `MutableByteArray` as buffers, add manual `keepAlive`