Before, on d38d85e537858be32555e0d13d85cf1af6319b67

```
Low-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.146791772s
IOPS:          228589

High-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.310599797s
IOPS:          200018
```

After, on bbeb81130ec3eafd8ced81564cc8bd46d24aff08

```
Low-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.157516626s
IOPS:          226471

High-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.315785114s
IOPS:          199230
```

Virtually the same