@jorisdral 2024-03-15
The latest commit added a guard that checks that byte arrays are pinned. There's probably a slight performance decrease, but not a very large one:

```
Low-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.163442451s
IOPS:          225318

High-level API benchmark
Total I/O ops: 262144
Elapsed time:  1.325703812s
IOPS:          197739
```